+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 ### Added
 
 - Support substring matcher for dyn filter, used when the query contains space. ([#411](https://github.com/liuchengxu/vim-clap/pull/411))
+- Add progress bar support for the download feature of maple. ([#419](https://github.com/liuchengxu/vim-clap/pull/419))
 
 ### Fixed